Refactor into pkg
