+++ pkg/helpers/struct.go
