+++ pkg/helpers/struct_test.go
